beauty_bot/requirements.txt
fastapi
uvicorn
requests
mysql-connector-python
python-multipart
jinja2